# docs/swagger.yaml
basePath: /api
definitions:
  endeus.CategoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Category'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
  model.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  utils.Error:
    properties:
      errors:
        additionalProperties: true
        type: object
      message: {}
      success:
        type: boolean
    type: object
...
host: 127.0.0.1:5000
info:
  contact: {}
  title: Swagger Endeus API
  version: "1.0"
paths:
  /category/all:
    get:
      consumes:
      - application/json
      description: Get all category
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/endeus.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
      summary: Get all category
      tags:
      - category
...
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"